Ice Fall + Fix document submission as directed by Canopy programmers. Turns out you have to shove some   crap on the the front of the base64 encoded pdf. Would have been nice if the docs said any thing   about it. + Handle doctor pagination correctly.

+++ src/main/java/org/oscarehr/integration/iceFall/service/transfer/IceFallCreatePrescriptionResponseTo1.java
@@ -43,7 +43,7 @@ public class IceFallCreatePrescriptionResponseTo1
 	@JsonDeserialize(using = LocalDateDeserializer.class)
 	private LocalDate registrationExpiry;
 	@JsonProperty("thclimit")
-	private Integer thcLimit;
+	private Float thcLimit;
 	private String diagnosis;
 	@JsonProperty("clinic")
 	private Integer clinicId;
@@ -90,12 +90,12 @@ public class IceFallCreatePrescriptionResponseTo1
 		this.registrationExpiry = registrationExpiry;
 	}
 
-	public Integer getThcLimit()
+	public Float getThcLimit()
 	{
 		return thcLimit;
 	}
 
-	public void setThcLimit(Integer thcLimit)
+	public void setThcLimit(Float thcLimit)
 	{
 		this.thcLimit = thcLimit;
 	}
